requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
apscheduler>=3.10.0
python-dotenv>=1.0.0
httpx>=0.26.0
python-telegram-bot>=21.0
boto3>=1.34.0
